Use lookahead to resolve `type` soft keyword (#11442)
## Summary

This PR updates the `type` alias statement parsing to use lookahead to
resolve whether it's used as a keyword or as an identifier depending on
the context.

The strategy here is that the first token should be either a name or a
soft keyword and the second can be either a `[` or `=`. Remember that
`type type = int` is valid where the first `type` is a keyword while the
second `type` is an identifier.