Update parser API references, expose `Program` to linter (#11505)
## Summary

This PR updates the parser API references per the updated function
signatures, especially the return type. Additionally, it exposes the
`Program` to the linter by using the new API and also storing it on the
`Checker` to be used by the rules.

Most of the changes are in test functions. There are only a handful of
usages in the non-test source code.