Replace `lex_starts_at` with `Tokens` in the formatter (#11515)
## Summary

This PR replaces the usage of `lex_starts_at` in the formatter with the
`Tokens` struct. This also updates the formatter API to take in the
`Program`. Earlier, it would take the individual parts of the program
but the formatter requires 3 fields from the program so we might as well
pass the containing struct itself.