Fix all compilation errors (#11612)
This PR fixes all the compilation error as raised by `clippy`. It
doesn't look at the tests. Apart from that it also does the following:

* Expose a `lex` method which creates the `Lexer` object. This is used
in the benchmark crate. Ideally, we shouldn't expose the lexer and in
the future we should move the lexer benchmark in the parser crate
* Update the parser reference in `red_knot` crate
* Change all references of `&[LexResult]` to `&Tokens`
* Add `CommentRanges` to the `LinterResult` as it's required after the
linter run. The `Program` is consumed because it requires an owned
`ParseError`