Use enum values instead of struct with single field (#11629)
## Summary

This PR is a follow-up to #11475 to use enum values where there's struct
with a single field. The main motivation to use struct was to keep the
snapshot update to a minimum which allows me to validate it. Now that
the validation is done, we can revert it back.

## Test Plan

Update the snapshots.