Make `ResolvedModule` a non-salsa tracked structk
A query result only needs to be a tracked struct if we intend to use it as a query ingredient.
It's unclear to me whether this is the case for `ResolvedModule`, that's why I make it a regular
struct for now. We can easily make it a tracked struct later on.